+++ src/RoutingControllersOptions.ts
@@ -9,7 +9,7 @@ import { CurrentUserChecker } from './CurrentUserChecker';
 export interface RoutingControllersOptions {
   /**
    * Indicates if cors are enabled.
-   * This requires installation of additional module (cors for express and kcors for koa).
+   * This requires installation of additional module (cors for express and @koa/cors for koa).
    */
   cors?: boolean | Object;