+++ test/functional/special-result-send.spec.ts
@@ -43,7 +43,9 @@ describe(``, () => {
       expressServer = createExpressServer().listen(3001, done);
     });
 
-    afterAll((done: DoneCallback) => expressServer.close(done));
+    afterAll((done: DoneCallback) => {
+      expressServer.close(done);
+    });
 
     it('should pipe stream to response', async () => {
       // expect.assertions(3);